+++ mex/editor/client/src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import type { Routes } from "@angular/router";
 import { StartPage } from "./start-page/start-page";
-import { FastTrackActivity } from "./fast-track-activity/fast-track-activity";
-import { FastTrackResource } from "./fast-track-resource/fast-track-resource";
+import { FastTrackActivity } from "./fast-track/fast-track-activity/fast-track-activity";
+import { FastTrackResource } from "./fast-track/fast-track-resource/fast-track-resource";
 
 /**
  * All registered Routes for the app.
