--- mex/editor/client/src/app/fast-track-resource/resource-submission.spec.ts
@@ -1,17 +1,0 @@
-import { TestBed } from "@angular/core/testing";
-
-import testProviders from "../../test-providers";
-import { ResourceSubmission } from "./resource-submission";
-
-describe("ResourceSubmission", () => {
-  let service: ResourceSubmission;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({ providers: [...testProviders] });
-    service = TestBed.inject(ResourceSubmission);
-  });
-
-  it("should be created", () => {
-    expect(service).toBeTruthy();
-  });
-});
