+++ mex/editor/client/src/app/fast-track/fast-track-activity/fast-track-activity.spec.ts
