+++ mex/editor/client/src/app/fast-track/fast-track-resource/concept-options.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from "@angular/core/testing";
+
+import { ConceptOptions } from "../../shared/concept-options.service";
+
+describe("VocabularySearchService", () => {
+  let service: ConceptOptions;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ConceptOptions);
+  });
+
+  it("should be created", () => {
+    expect(service).toBeTruthy();
+  });
+});
