+++ mex/editor/client/src/app/fast-track/fast-track-resource/contact-list/contact-list.scss
