+++ mex/editor/client/src/app/fast-track/fast-track-resource/fast-track-resource.scss
