+++ mex/editor/client/src/app/fast-track/fieldset/fieldset.spec.ts
@@ -2,6 +2,7 @@ import type { ComponentFixture } from "@angular/core/testing";
 import { TestBed } from "@angular/core/testing";
 
 import { Fieldset } from "./fieldset";
+import { inputBinding } from "@angular/core";
 
 describe("Fieldset", () => {
   let component: Fieldset<unknown>;
@@ -12,7 +13,9 @@ describe("Fieldset", () => {
       imports: [Fieldset],
     }).compileComponents();
 
-    fixture = TestBed.createComponent(Fieldset);
+    fixture = TestBed.createComponent(Fieldset, {
+      bindings: [inputBinding("labelKey", () => "test.fieldset.label")],
+    });
     component = fixture.componentInstance;
     await fixture.whenStable();
   });
